refactor: store surcharge rates in GasLedger

+++ src/contracts/interfaces/IDAppControl.sol
@@ -24,7 +24,7 @@ interface IDAppControl {
 
     function getDAppGasLimit() external view returns (uint32);
 
-    function getBundlerSurchargeRate() external view returns (uint128);
+    function getBundlerSurchargeRate() external view returns (uint24);
 
     function getBidFormat(UserOperation calldata userOp) external view returns (address bidToken);
